LayoutTests/platform/wk2/http/tests/security/blocked-on-redirect-expected.txt

CONSOLE MESSAGE: Not allowed to use restricted network port 25: http://localhost:25/security/resources/post-done.html
This test produces output in the console. Only the initial URL should be logged.
http://127.0.0.1:8000/security/blocked-on-redirect.html - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/security/blocked-on-redirect.html, main document URL http://127.0.0.1:8000/security/blocked-on-redirect.html, http method GET> redirectResponse (null)
http://127.0.0.1:8000/security/blocked-on-redirect.html - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/security/blocked-on-redirect.html, http status code 200>
http://127.0.0.1:8000/security/blocked-on-redirect.html - didFinishLoading
http://127.0.0.1:8000/resources/redirect.py?code=307&url=http://localhost:25/security/resources/post-done.html - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/resources/redirect.py?code=307&url=http://localhost:25/security/resources/post-done.html, main document URL http://127.0.0.1:8000/security/blocked-on-redirect.html, http method GET> redirectResponse (null)
